use one topsuite per module.
