+++ internal/health/components_int_test.go
@@ -2,44 +2,26 @@ package health
 
 import (
 	"context"
-	"testing"
 
 	"github.com/G-Research/unicorn-history-server/internal/yunikorn"
 	"github.com/G-Research/unicorn-history-server/test/config"
-	"github.com/G-Research/unicorn-history-server/test/database"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
 
 type ComponentsTestSuite struct {
 	suite.Suite
-	tp             *database.TestPostgresContainer
 	pool           *pgxpool.Pool
 	yunikornClient *yunikorn.RESTClient
 }
 
 func (ts *ComponentsTestSuite) SetupSuite() {
-	ctx := context.Background()
-	cfg := database.InstanceConfig{
-		User:     "test",
-		Password: "test",
-		DBName:   "template",
-		Host:     "localhost",
-		Port:     15436,
-	}
-
-	tp, err := database.NewTestPostgresContainer(ctx, cfg)
-	require.NoError(ts.T(), err)
-	ts.tp = tp
-	ts.pool = tp.Pool(ctx, ts.T(), &cfg)
 	ts.yunikornClient = yunikorn.NewRESTClient(config.GetTestYunikornConfig())
 }
 
 func (ts *ComponentsTestSuite) TearDownSuite() {
-	err := ts.tp.Container.Terminate(context.Background())
-	require.NoError(ts.T(), err)
+	ts.pool.Close()
 }
 
 func (ts *ComponentsTestSuite) TestNewComponents() {
@@ -72,11 +54,4 @@ func (ts *ComponentsTestSuite) TestNewComponents() {
 			assert.Equal(ts.T(), tt.expectedHealthy, status.Healthy)
 		})
 	}
-}
-
-func TestComponentsIntegrationTestSuite(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping integration test in short mode.")
-	}
-	suite.Run(t, new(ComponentsTestSuite))
 }
